requirements.txt
osmnx>=1.3.0
matplotlib>=3.7.0
geopandas>=0.14.0
pillow>=10.0.0
